requirements.txt
jax==0.4.28
jaxlib==0.4.28
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
numpy==1.26.4
flax
chex 
jaxtyping
dm-haiku
optax
hydra-core 
omegaconf
scipy
sbi 
sbibm 
matplotlib 
seaborn 
pandas 
tueplots
pytest
numpyro
pyro-ppl
